Split Tauri command wrappers
