+++ src/api/tauri-commands/types.ts
@@ -1,0 +1,24 @@
+// Re-export types so existing consumers don't break.
+export type {
+  AccountDto,
+  AccountSyncStatusDto,
+  AppError,
+  ArticleDto,
+  BrowserWebviewState,
+  DatabaseInfoDto,
+  DevRuntimeOptions,
+  DiscoveredFeedDto,
+  FeedArticleSummaryDto,
+  FeedDto,
+  FeedIntegrityCleanupDto,
+  FeedIntegrityReportDto,
+  FolderDto,
+  MuteKeywordDto,
+  OldUnreadDays,
+  OldUnreadScopeKind,
+  PlatformInfo,
+  PlatformPermissionDeniedRecovery,
+  PreferencesDto,
+  TagDto,
+  UpdateInfoDto,
+} from "@/api/schemas";
